refactor: refactor mypage #108
